cmse: enable SAU defines for non-secure state
The SAU register defines and bitfields are still useful for applications
that are not running in the secure state.

e.g. An application triggers a SecureFault, during which the secure
handler preserved the `SFSR` register value in `.noinit` memory. Once
rebooted, the non-secure application queries the `SFSR` register and
wishes to inspect the bits to determine more precisely the cause.

Signed-off-by: Jordan Yates <jordan@embeint.com>